requirements.txt
Django
pytz
psycopg2
pillow
django-extensions
django-model-utils
django-countries
djangorestframework
djangorestframework-camel-case
